Move service/builder to internal, no need to expose public (#2625)
