+++ service/internal/builder/testdata/bad_processor_factory.yaml
