+++ service/internal/builder/testdata/bad_receiver_factory.yaml
