+++ service/internal/builder/testdata/pipelines_builder.yaml
