+++ service/internal/builder/testdata/unused_receiver.yaml
